Move cypress tests to ws-nextjs-app
