+++ ws-nextjs-app/cypress/e2e/liveRadioPage/tests.ts
@@ -1,7 +1,6 @@
 /* eslint-disable import/prefer-default-export */
 
-import path from 'ramda/src/path';
-import chartbeatTests from '../../../support/helpers/chartbeatTests';
+import chartbeatTests from '../../support/helpers/chartbeatTests';
 // For testing features that may differ across services but share a common logic e.g. translated strings.
 export default ({ service, pageType }) =>
   describe(`Tests for ${service} ${pageType}`, () => {
@@ -34,10 +33,7 @@ export default ({ service, pageType }) =>
     describe('Radio Schedule', () => {
       it('should be displayed if there is enough schedule data', function test() {
         cy.fixture(`toggles/${service}.json`).then(toggles => {
-          const scheduleIsEnabled = path(
-            ['liveRadioSchedule', 'enabled'],
-            toggles,
-          );
+          const scheduleIsEnabled = toggles?.liveRadioSchedule?.enabled;
           cy.log(
             `Live Radio Page configured for Radio Schedule? ${scheduleIsEnabled}`,
           );
